pyproject.toml
[tool.poetry]
name = "mooose-bcf"
version = "0.1.0"
description = ""
authors = ["Murray Tait <murray.tait@infinityworks.com>"]

[tool.poetry.dependencies]
python = "~3.8"
pandas = "^1.4.1"
matplotlib = "^3.5.1"
google-cloud = "^0.34.0"
google-cloud-storage = "^2.1.0"
google-cloud-bigquery = "^2.34.0"
google-cloud-functions = "^1.5.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
pytest-html = "^3.1.1"
mock = "^4.0.3"
flake8 = "^4.0.1"
black = "^22.1.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
